% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coordinates.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Get Coordinates In Decimal Form}
\usage{
get_coordinates(latitudeList, longitudeList, delimitor)
}
\arguments{
\item{latitudeList}{a list of latitudes in "xx°yy.zz'N/S/E/W" form}

\item{longitudeList}{a list of longitudes in "xx°yy.zz'N/S/E/W" form}

\item{delimiter}{a delimiter that separates the lat and long coordinates in the final list}
}
\description{
"get_coordinates" takes a list of latitudes and a list of longitudes in degree/minute format and returns a list of concatenated coordinates in decimal form separated by a delimiter of the user's choice.
}
\examples{
get_coordinates("55°44.23'N", "78°23.5'W", ", " )
}
\keyword{coordinates}
\keyword{east}
\keyword{latitude}
\keyword{longitude}
\keyword{north}
\keyword{south}
\keyword{west}
